+++ rust/main/chains/hyperlane-ethereum/src/rpc_clients/provider.rs
@@ -6,7 +6,7 @@ use std::time::Duration;
 
 use async_trait::async_trait;
 use derive_new::new;
-use ethers::types::{Block, H256 as EthersH256};
+use ethers::types::{Block, H160, H256 as EthersH256};
 use ethers::{prelude::Middleware, types::TransactionReceipt};
 use ethers_contract::builders::ContractCall;
 use ethers_core::abi::Function;
@@ -87,7 +87,7 @@ where
 
 /// Methods of provider which are used in submitter
 #[async_trait]
-pub trait EvmProviderForSubmitter: Send + Sync {
+pub trait EvmProviderForLander: Send + Sync {
     /// Get the transaction receipt for a given transaction hash
     async fn get_transaction_receipt(
         &self,
@@ -134,11 +134,11 @@ pub trait EvmProviderForSubmitter: Send + Sync {
     ) -> ChainResult<ZksyncEstimateFeeResponse>;
 
     /// Get default sender
-    fn default_sender(&self) -> Option<Address>;
+    fn get_signer(&self) -> Option<H160>;
 }
 
 #[async_trait]
-impl<M> EvmProviderForSubmitter for EthereumProvider<M>
+impl<M> EvmProviderForLander for EthereumProvider<M>
 where
     M: Middleware + 'static,
 {
@@ -231,7 +231,7 @@ where
             .map_err(ChainCommunicationError::from_other)
     }
 
-    fn default_sender(&self) -> Option<Address> {
+    fn get_signer(&self) -> Option<H160> {
         self.provider.default_sender()
     }
 }
@@ -350,7 +350,7 @@ where
                 ChainCommunicationError::Other(HyperlaneCustomErrorWrapper::new(Box::new(e)))
             })?
         else {
-            tracing::trace!(domain=?self.domain, "Latest block not found");
+            tracing::trace!(domain=?self.domain.name(), "Latest block not found");
             return Ok(None);
         };
 
@@ -394,7 +394,7 @@ pub struct SubmitterProviderBuilder {}
 
 #[async_trait]
 impl BuildableWithProvider for SubmitterProviderBuilder {
-    type Output = Box<dyn EvmProviderForSubmitter>;
+    type Output = Box<dyn EvmProviderForLander>;
     const NEEDS_SIGNER: bool = true;
 
     // the submitter does not use the ethers submission middleware.