+++ typescript/cli/src/commands/warp.ts
@@ -113,7 +113,12 @@ export const apply: CommandModuleWithWarpApplyContext<{
       coerce: (dir) => removeEndingSlash(dir),
     },
   },
-  handler: async ({ context, strategy: strategyUrl, receiptsDir }) => {
+  handler: async ({
+    context,
+    strategy: strategyUrl,
+    receiptsDir,
+    warpRouteId,
+  }) => {
     logCommandHeader('Hyperlane Warp Apply');
 
     if (strategyUrl)
@@ -126,6 +131,7 @@ export const apply: CommandModuleWithWarpApplyContext<{
       warpCoreConfig: context.warpCoreConfig,
       strategyUrl,
       receiptsDir,
+      warpRouteId,
     });
     process.exit(0);
   },
@@ -150,7 +156,7 @@ export const deploy: CommandModuleWithWarpDeployContext<{
     },
     warpRouteId: warpRouteIdCommandOption,
   },
-  handler: async ({ context, dryRun }) => {
+  handler: async ({ context, dryRun, warpRouteId }) => {
     logCommandHeader(
       `Hyperlane Warp Route Deployment${dryRun ? ' Dry-Run' : ''}`,
     );
@@ -160,6 +166,7 @@ export const deploy: CommandModuleWithWarpDeployContext<{
         context,
         // Already fetched in the resolveWarpRouteConfigChains
         warpDeployConfig: context.warpDeployConfig,
+        warpRouteId,
       });
     } catch (error: any) {
       evaluateIfDryRunFailure(error, dryRun);