+++ typescript/sdk/src/token/deploy.ts
@@ -33,6 +33,7 @@ import { GasRouterDeployer } from '../router/GasRouterDeployer.js';
 import { resolveRouterMapConfig } from '../router/types.js';
 import { ChainMap, ChainName } from '../types.js';
 
+import { TokenMetadataMap } from './TokenMetadataMap.js';
 import { TokenType, gasOverhead } from './config.js';
 import {
   HypERC20Factories,
@@ -48,7 +49,6 @@ import {
   CctpTokenConfig,
   HypTokenConfig,
   HypTokenRouterConfig,
-  TokenMetadata,
   TokenMetadataSchema,
   WarpRouteDeployConfig,
   WarpRouteDeployConfigMailboxRequired,
@@ -113,7 +113,7 @@ abstract class TokenDeployer<
     loggerName: string,
     ismFactory?: HyperlaneIsmFactory,
     contractVerifier?: ContractVerifier,
-    concurrentDeploy = false,
+    concurrentDeploy = true,
   ) {
     super(multiProvider, factories, {
       logger: rootLogger.child({ module: loggerName }),
@@ -203,10 +203,20 @@ abstract class TokenDeployer<
   static async deriveTokenMetadata(
     multiProvider: MultiProvider,
     configMap: WarpRouteDeployConfig,
-  ): Promise<TokenMetadata | undefined> {
-    for (const [chain, config] of Object.entries(configMap)) {
+  ): Promise<TokenMetadataMap> {
+    const metadataMap = new TokenMetadataMap();
+
+    const priorityGetter = (type: string) => {
+      return ['collateral', 'native'].indexOf(type);
+    };
+
+    const sortedEntries = Object.entries(configMap).sort(
+      ([, a], [, b]) => priorityGetter(b.type) - priorityGetter(a.type),
+    );
+
+    for (const [chain, config] of sortedEntries) {
       if (isTokenMetadata(config)) {
-        return TokenMetadataSchema.parse(config);
+        metadataMap.set(chain, TokenMetadataSchema.parse(config));
       } else if (multiProvider.getProtocol(chain) !== ProtocolType.Ethereum) {
         // If the config didn't specify the token metadata, we can only now
         // derive it for Ethereum chains. So here we skip non-Ethereum chains.
@@ -216,9 +226,13 @@ abstract class TokenDeployer<
       if (isNativeTokenConfig(config)) {
         const nativeToken = multiProvider.getChainMetadata(chain).nativeToken;
         if (nativeToken) {
-          return TokenMetadataSchema.parse({
-            ...nativeToken,
-          });
+          metadataMap.set(
+            chain,
+            TokenMetadataSchema.parse({
+              ...nativeToken,
+            }),
+          );
+          continue;
         }
       }
 
@@ -238,10 +252,14 @@ abstract class TokenDeployer<
             erc721.name(),
             erc721.symbol(),
           ]);
-          return TokenMetadataSchema.parse({
-            name,
-            symbol,
-          });
+          metadataMap.set(
+            chain,
+            TokenMetadataSchema.parse({
+              name,
+              symbol,
+            }),
+          );
+          continue;
         }
 
         let token: string;
@@ -270,15 +288,19 @@ abstract class TokenDeployer<
           erc20.decimals(),
         ]);
 
-        return TokenMetadataSchema.parse({
-          name,
-          symbol,
-          decimals,
-        });
+        metadataMap.set(
+          chain,
+          TokenMetadataSchema.parse({
+            name,
+            symbol,
+            decimals,
+          }),
+        );
       }
     }
 
-    return undefined;
+    metadataMap.finalize();
+    return metadataMap;
   }
 
   protected async configureCctpDomains(
@@ -440,9 +462,9 @@ abstract class TokenDeployer<
   }
 
   async deploy(configMap: WarpRouteDeployConfigMailboxRequired) {
-    let tokenMetadata: TokenMetadata | undefined;
+    let tokenMetadataMap: TokenMetadataMap;
     try {
-      tokenMetadata = await TokenDeployer.deriveTokenMetadata(
+      tokenMetadataMap = await TokenDeployer.deriveTokenMetadata(
         this.multiProvider,
         configMap,
       );
@@ -451,8 +473,13 @@ abstract class TokenDeployer<
       throw err;
     }
 
-    const resolvedConfigMap = objMap(configMap, (_, config) => ({
-      ...tokenMetadata,
+    const resolvedConfigMap = objMap(configMap, (chain, config) => ({
+      name: tokenMetadataMap.getName(chain),
+      decimals: tokenMetadataMap.getDecimals(chain),
+      symbol:
+        tokenMetadataMap.getSymbol(chain) ||
+        tokenMetadataMap.getDefaultSymbol(),
+      scale: tokenMetadataMap.getScale(chain),
       gas: gasOverhead(config.type),
       ...config,
     }));
@@ -476,7 +503,7 @@ export class HypERC20Deployer extends TokenDeployer<HypERC20Factories> {
     multiProvider: MultiProvider,
     ismFactory?: HyperlaneIsmFactory,
     contractVerifier?: ContractVerifier,
-    concurrentDeploy = false,
+    concurrentDeploy = true,
   ) {
     super(
       multiProvider,