+++ typescript/github-proxy/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hyperlane-xyz/github-proxy
 
+## 13.2.1
+
+## 13.2.0
+
+## 13.1.1
+
+### Patch Changes
+
+- ba4deea: Revert workspace dependency syntax.
+
+## 13.1.0
+
+## 13.0.0
+
 ## 12.6.0
 
 ## 12.5.0